thorium_node: use 64-bit integers for counters to avoid overflow

+++ engine/thorium/node.h
@@ -220,8 +220,8 @@ struct thorium_node {
     struct core_counter counter;
 #endif
 
-    int counter_last_received_message_count;
-    int counter_last_sent_message_count;
+    uint64_t counter_last_received_message_count;
+    uint64_t counter_last_sent_message_count;
 
     /*
      * Requires -D_POSIX_C_SOURCE=200112L
